requirements.txt
python-telegram-bot>=20.0
python-dotenv>=0.19.0
psutil>=5.8.0
rich>=10.0.0
asyncio>=3.4.3
